chore: remove old sourcemap search strategies
The turbo predictor will have been default in stable for 6 months, let's remove the old search strategies
