+++ src/ioc.ts
@@ -58,10 +58,9 @@ import { IFsUtils, LocalAndRemoteFsUtils, LocalFsUtils } from './common/fsUtils'
 import { ILogger } from './common/logging';
 import { Logger } from './common/logging/logger';
 import { createMutableLaunchConfig, MutableLaunchConfig } from './common/mutableLaunchConfig';
-import { CodeSearchStrategy } from './common/sourceMaps/codeSearchStrategy';
 import { IRenameProvider, RenameProvider } from './common/sourceMaps/renameProvider';
 import { CachingSourceMapFactory, ISourceMapFactory } from './common/sourceMaps/sourceMapFactory';
-import { ISearchStrategy, ISearchStrategyFallback } from './common/sourceMaps/sourceMapRepository';
+import { ISearchStrategy } from './common/sourceMaps/sourceMapRepository';
 import { TurboSearchStrategy } from './common/sourceMaps/turboSearchStrategy';
 import { ISourcePathResolver } from './common/sourcePathResolver';
 import { AnyLaunchConfiguration } from './configuration';
@@ -230,10 +229,6 @@ export const createTopLevelSessionContainer = (parent: Container) => {
   container.bind(OutFiles).to(OutFiles).inSingletonScope();
   container.bind(VueComponentPaths).to(VueComponentPaths).inSingletonScope();
   container.bind(IVueFileMapper).to(VueFileMapper).inSingletonScope();
-  container
-    .bind(ISearchStrategyFallback)
-    .toDynamicValue(ctx => CodeSearchStrategy.createOrFallback(ctx.container.get<ILogger>(ILogger)))
-    .inSingletonScope();
   container.bind(ISearchStrategy).to(TurboSearchStrategy).inSingletonScope();
 
   container.bind(INodeBinaryProvider).to(InteractiveNodeBinaryProvider);
